+++ NEWS.md
@@ -13,6 +13,7 @@ NEW FEATURES :
 * `createClusterST()`/`editClusterST()` :  
   - **New properties** (*efficiencywithdrawal*, *penalize-variation-injection*, *penalize-variation-withdrawal*, see list of properties according to study version of Antares with `storage_values_default()`)  
   - **New optional time series** (cost-injection, cost-withdrawal, cost-level, cost-variation-injection, cost-variation-withdrawal)
+* `updateScenarioBuilder()` New type of series "hfl" ("hydro final level", similar to "hydrolevels") is available   
 
 
 ### Breaking changes  :